requirements.txt
fastText == 0.9.2
mecab-python3 == 1.0.4
requests == 2.18
neologdn ==0.5.1